feat: implement custom processor actions for all facets that involve calculating the length of a pcommon.Slice (#88)
OTTL's `Len()` doesn't work well with some of the slices obtained by
unmarshalling k8s events into their respective API entities. This is
because `Len()` doesn't support the type of the inner elements of most
of the slices we have to use.
As a result, the only way to "calculate" such lengths is via custom code
in our events' processor.