fix: OB-37389 Obfuscate secrets in YAML body (#101)
Add whole processor framework to modify an event body in place. The set
of actions that do that are also type-specific, meaning that we only run
relevant actions on any given event, based on the object type.

These actions modify the unmarshalled object IN PLACE and BEFORE the
attributes-computing actions do their thing. This is mainly to allow
secrets' obfuscation, where we want to prevent secret stuff to end up in
the attributes by mistake.

Add secret "body" actions that obfuscate the secrets in "data".

### Description

OB-XXX Please explain the changes you made here.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary