fix: Override log body with structured object
Currently we were modifying the marshalled k8s object and then
overriding the log body with its unmarshalled string representation.

This doesn't work well when the custom processor is not last in the
pipeline, because other processors might have to index into the
structured body to be able to extract values.

Since we need to move the transform processor after the custom processor
in the agent configmap, we want to make sure that this processor leaves
the log body as a structured object, rather than a plain string.