fix: clarify and expand the default env vars the agent provides to the configs (#142)
### Description

Clarify and expand the default env vars the agent provides to the
configs. This makes the vanilla collector endpoint available as well as
exposing the prometheus endpoint. It also renames OBSERVE_TOKEN to
OBSERVE_AUTHORIZATION_HEADER for clarity since the value is prefixed by
`"Bearer "`, and renames OBSERVE_ENDPOINT to OBSERVE_OTEL_ENDPOINT since
the value is suffixed with `"/v2/otel"`

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary