feat: improve internal_telemetry agent configuration (#191)
### Description

Improve internal_telemetry agent configuration. Currently the only way
in the agent to set logging to a level other than INFO or DEBUG is with
custom otel config overrides. This allows for easy configuration via the
agent config as well as respecting the env var `OTEL_LOG_LEVEL` if it's
set outside the agent process.